backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# MongoDB
beanie==1.24.0
motor==3.3.2

# WebSocket
websockets==12.0

# Utils
python-dateutil==2.8.2
aiohttp==3.9.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
coverage==7.4.1